% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoc.R
\name{solve_all}
\alias{solve_all}
\title{Solve all problems using input files from a directory}
\usage{
solve_all(dir)
}
\arguments{
\item{dir}{A directory of input files (including day number in name).}
}
\description{
Solve all problems using input files from a directory
}
