% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoc.R
\name{solve_day}
\alias{solve_day}
\title{Solve a day's problem for both parts}
\usage{
solve_day(file, day = guess_day(file))
}
\arguments{
\item{day}{The day to solve.}

\item{input}{An input file.}
}
\description{
Solve a day's problem for both parts
}
