% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoc.R
\name{solve_day}
\alias{solve_day}
\title{Solve a day's problem for both parts}
\usage{
solve_day(file = NULL, day = NULL, input = NULL)
}
\arguments{
\item{file}{An input file.}

\item{day}{The day to solve.}

\item{input}{A direct input (numeric or string).}
}
\description{
Solve a day's problem for both parts
}
